requirements.txt
django==4.0.6
python-telegram-bot==13.13
pillow==9.2.0
django-hashedfilenamestorage=2.4.1
